src/main/java/guru/springframework/services/springdata/CategoryServiceSDImpl.java
package guru.springframework.services.springdata;

import guru.springframework.model.Category;
import guru.springframework.repositories.CategoryRepository;
import guru.springframework.services.CategoryService;
import org.springframework.stereotype.Service;

import java.util.HashSet;
import java.util.Set;

@Service
public class CategoryServiceSDImpl implements CategoryService {

    private CategoryRepository categoryRepository;

    public CategoryServiceSDImpl(CategoryRepository categoryRepository) {
        this.categoryRepository = categoryRepository;
    }

    @Override
    public Category findById(Long id) {
        return categoryRepository.findById(id).orElse(null);
    }

    @Override
    public Set<Category> findAll() {
        Set<Category> categories = new HashSet<>();
        categoryRepository.findAll().forEach(categories::add);
        System.out.println(categories.size());
        return categories;
    }

    @Override
    public Category save(Category category) {
        return categoryRepository.save(category);
    }

    @Override
    public void delete(Category category) {
        categoryRepository.delete(category);
    }

    @Override
    public void deleteById(Long id) {
        categoryRepository.deleteById(id);
    }
}
